services/math-grader/math_grading_module/requirements.txt
flask
flask-cors
requests
python-dotenv
langchain-core
